+++ HISTORY.rst
@@ -32,6 +32,7 @@ latest (unreleased)
 * Migrate addon to add a wizard to split products in multiples packs
 * Migrate Swiss PP labels addon
 * Migrate addon to select a logo per shop on postlogistics delivery labels
+* Migrate Swiss PP labels addon
 
 **Bugfixes**
 * Analytic accounts : allow to "search more..." on SO